+++ node/src/types/kernel/api/convert/pandoc/index.d.ts
@@ -15,16 +15,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IResponse } from "@/types/kernel";
-import * as axios from "axios";
-
-export class KernelError extends Error {
-    public readonly code: number;
-    constructor(
-        public readonly body: IResponse,
-        public readonly response: axios.AxiosResponse | Response,
-    ) {
-        super(body.msg);
-        this.code = body.code;
-    }
-}
+export { IPayload } from "./payload";
+export { IResponse } from "./response";