+++ node/src/types/kernel/api/convert/pandoc/response.d.ts
@@ -15,16 +15,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IResponse } from "@/types/kernel";
-import * as axios from "axios";
-
-export class KernelError extends Error {
-    public readonly code: number;
-    constructor(
-        public readonly body: IResponse,
-        public readonly response: axios.AxiosResponse | Response,
-    ) {
-        super(body.msg);
-        this.code = body.code;
-    }
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    data: null;
+    /**
+     * status message
+     */
+    msg: string;
 }